Update mocking for all server tests and ensure no broken tests

+++ server/controllers/userController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import User from '../models/userModel';
 import admin from '../firebase/config';
+import usingAuth from "../usingAuth";
 
 // User Profile API
 
@@ -88,7 +89,9 @@ const updateUser = async (req: Request, res: Response) => {
 
 const deleteUser = async (req: Request, res: Response) => {
   try {
-    await admin.auth().deleteUser(req.params.user_id);
+    if (usingAuth()) {
+      await admin.auth().deleteUser(req.params.user_id);
+    }
     const user = await User.findOneAndDelete({ user_id: req.params.user_id });
     res.status(200).json({
       status: 'success',
@@ -184,7 +187,7 @@ const acceptConnectionRequest = async (req: Request, res: Response) => {
     }
     return res.status(403).json({
       status: 'failure',
-      message: 'Error accepting connection request',
+      message: 'No incoming request to accept',
     });
   } catch (err) {
     res.status(400).json({
